requirements.txt
-i https://pypi.tuna.tsinghua.edu.cn/simple

allure-pytest==2.11.1
Faker==15.1.1
pandas==1.5.3
openpyxl==3.1.0
PyMySQL==1.0.2
pytest==7.2.0
pytest-ordering==0.6
pytest-rerunfailures==11.1
PyYAML==6.0
requests==2.28.1
click==8.1.3
urllib3==1.26.12
python-jenkins==1.7.0
loguru==0.6.0
ddddocr==1.4.7
dingtalkchatbot==1.5.7
pydantic==1.10.5
jsonpath~=0.82